+++ kifi-backend/modules/eliza/app/com/keepit/eliza/model/UriRenormalizationTrackingRepo.scala
@@ -13,40 +13,18 @@ import com.google.inject.{Inject, Singleton, ImplementedBy}
 
 import org.joda.time.DateTime
 
-
-
-case class UriRenormalizationEvent(
-    id: Option[Id[UriRenormalizationEvent]] = None,
-    createdAt: DateTime = currentDateTime,
-    updateAt: DateTime = currentDateTime,
-    sequenceNumber: Long,
-    numIdsChanged: Long,
-    idsRetired: Seq[Id[NormalizedURI]]
-  )
-  extends Model[UriRenormalizationEvent] {
-
-  def withId(id: Id[UriRenormalizationEvent]): UriRenormalizationEvent = this.copy(id = Some(id))
-  def withUpdateTime(updateTime: DateTime) = this.copy(updateAt=updateTime)
-}
-
-
-
-
-
 @ImplementedBy(classOf[UriRenormalizationTrackingRepoImpl])
 trait UriRenormalizationTrackingRepo extends Repo[UriRenormalizationEvent] {
 
   def getCurrentSequenceNumber()(implicit session: RSession): Long
 
   def addNew(sequenceNumber: Long, numIdsChanged: Long, idsRetired: Seq[Id[NormalizedURI]])(implicit session: RWSession) : Unit
-
 }
 
 class UriRenormalizationTrackingRepoImpl @Inject() (
     val clock: Clock,
     val db: DataBaseComponent
   ) extends DbRepo[UriRenormalizationEvent] with UriRenormalizationTrackingRepo {
-
 
   override val table = new RepoTable[UriRenormalizationEvent](db, "uri_renormalization_event") {
     def sequenceNumber = column[Long]("sequence_number", O.NotNull)
