+++ kifi-backend/modules/eliza/app/com/keepit/eliza/model/UserThreadRepo.scala
@@ -25,47 +25,6 @@ import scala.concurrent.duration._
 import MessagingTypeMappers._
 import com.keepit.common.mail.{PostOffice, ElectronicMailCategory}
 
-case class Notification(thread: Id[MessageThread], message: Id[Message])
-case class NotificationCategory(category: String)
-object NotificationCategory {
-  implicit def fromElectronicMailCategory(emailCategory: ElectronicMailCategory): NotificationCategory = NotificationCategory(emailCategory.category)
-  object Global {
-    val ANNOUNCEMENT = NotificationCategory("ANNOUNCEMENT")
-  }
-
-  object Personal {
-    val MESSAGE = fromElectronicMailCategory(PostOffice.Categories.User.MESSAGE)
-  }
-}
-
-case class UserThreadActivity(id: Id[UserThread], threadId: Id[MessageThread], userId: Id[User], lastActive: Option[DateTime], started: Boolean, lastSeen: Option[DateTime])
-
-
-case class UserThread(
-    id: Option[Id[UserThread]] = None,
-    createdAt: DateTime = currentDateTime,
-    updateAt: DateTime = currentDateTime,
-    user: Id[User],
-    thread: Id[MessageThread],
-    uriId: Option[Id[NormalizedURI]],
-    lastSeen: Option[DateTime],
-    unread: Boolean = false,
-    muted: Boolean = false,
-    lastMsgFromOther: Option[Id[Message]],
-    lastNotification: JsValue,
-    notificationUpdatedAt: DateTime = currentDateTime,
-    notificationLastSeen: Option[DateTime] = None,
-    notificationEmailed: Boolean = false,
-    replyable: Boolean = true,
-    lastActive: Option[DateTime] = None, //Contains the 'createdAt' timestamp of the last message this user sent on this thread
-    started: Boolean = false //Wether or not this thread was started by this user
-  )
-  extends Model[UserThread] {
-
-  def withId(id: Id[UserThread]): UserThread = this.copy(id = Some(id))
-  def withUpdateTime(updateTime: DateTime) = this.copy(updateAt=updateTime)
-}
-
 
 @ImplementedBy(classOf[UserThreadRepoImpl])
 trait UserThreadRepo extends Repo[UserThread] {
